Particle package refractor + feather , testing necromancer in raids. Fix warlock armor dye recipe not working

+++ src/main/java/elucent/eidolon/recipe/ItemRitualRecipe.java
@@ -130,7 +130,7 @@ public class ItemRitualRecipe extends RitualRecipe {
                 }
             }
 
-            int healthRequirement = buffer.readInt();
+            float healthRequirement = buffer.readFloat();
 
             ItemStack output = buffer.readItem();
             boolean keepNbtOfReagent = buffer.readBoolean();
